pyproject.toml
[tool.poetry]
name = "twiskey"
version = "0.1.0"
description = "TwitterとMisskeyに同時投稿出来るコマンドです。"
authors = ["Comamoca <comamoca.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.10.0"
"Misskey.py" = "^4.1.0"
python-dotenv = "^0.20.0"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
